If available, use three-way-compare with iterator_facade:
This also fixes an issue with MSVC name lookup:
If the operand of a function or operator call is of type S, and S is
a nested class of D, and D inherits from B, then the friend functions
and operators defined in B will not be found in name lookup unless
D is the first operand.

This occurs when the sentinel_type of an iterator is used on the left
hand of a comparison operator, and the sentinel_type is a nested
class of that iterator.